Macro librustlet::binary

macro_rules! binary {
    () => { ... };
    ($a : expr, $b : expr) => { ... };
}

The binary macro is used to send and receive binary websocket messages. If no parameters are specified, this macro will return the payload of the message that is received from a Socklet::Binary event. The payload, which is a Vec is wrapped in a std::result::Result enum and an error may be returned in the event of a misconfiguration. If two parameters are specified, this macro will write data (second parameter) to the handle (first parameter). The data is of type std::slice.

Also see

rustlet_init

socklet

socklet_mapping

handle

event

text

ping

pong

Examples

use nioruntime_err::Error;
use librustlet::*;
use nioruntime_log::*;

debug!();

fn test() -> Result<(), Error> {
    rustlet_init!(
        RustletConfig {
            http_config: HttpConfig {
                port: 80,
                ..HttpConfig::default()
            },
            ..RustletConfig::default()
        }
    );

    socklet!("mysocklet", {
        let handle = handle!()?;
        let id = handle.get_connection_id();
        match event!()? {
            Socklet::Binary => {
                let binary = binary!()?;
                info!("got binary on [cid={}]: {:?}", id, binary);
                let bin_resp = &[1,2,3,4,5];
                binary!(handle, bin_resp);
            }
            _ => {},
        }
    });

    socklet_mapping!("/mysocklet", "mysocklet");

    Ok(())
}