Macro librustlet::pong

macro_rules! pong {
    ($a : expr) => { ... };
}

Send a websocket pong message. See https://datatracker.ietf.org/doc/html/rfc6455 for more details on the websocket protocol and the usage of pongs. Note that most browsers do not support this message type.

Also see

rustlet_init

socklet

socklet_mapping

handle

event

text

ping

binary

Examples

use nioruntime_err::Error;
use librustlet::*;
use nioruntime_log::*;

debug!();

fn test() -> Result<(), Error> {
    rustlet_init!(
        RustletConfig {
            http_config: HttpConfig {
                port: 80,
                ..HttpConfig::default()
            },
            ..RustletConfig::default()
        }
    );

    socklet!("mysocklet", {
        let handle = handle!()?;
        let id = handle.get_connection_id();
        match event!()? {
            Socklet::Text => {
                let text = text!()?;
                info!("got text [cid={}]: {}", id, text);
                text!(handle, "echo [cid={}]: '{}'", id, text);
                ping!(handle);
            },
            Socklet::Ping => {
                pong!(handle);
            },
            _ => {},
        }
    });

    socklet_mapping!("/mysocklet", "mysocklet");

    Ok(())
}