Macro librustlet::rustlet

macro_rules! rustlet {
    ($a:expr,$b:expr) => { ... };
}

Specifies a rustlet. Rustlets are closures that process HTTP requests and generate a response, so variables can be moved into them and shared among other rustlets or any other closure. Rustlets are processed in the nioruntime. So, the exectuion is performant. See the other macros for detailed examples on how to use all of the functionality of rustlets. To try out a minimally simple rustlet as a project, see rustlet-simple.

Also see

add_header

async_complete

async_context

cookie

flush

header_len

header_name

header_value

query

request

request_content

response

rustlet_init

rustlet_mapping

session

session_delete

set_content_type

set_cookie

set_redirect

Examples

use nioruntime_util::Error;
use librustlet::*;
use std::sync::{Mutex, Arc};
use nioruntime_log::*;

debug!();

fn test() -> Result<(), Error> {
    let count = Arc::new(Mutex::new(0));
    let count_clone = count.clone();

    // init the rustlet container, in this case with default values
    rustlet_init!(RustletConfig::default());

    // define our first rustlet
    rustlet!("myrustlet", {
        let mut count = count.lock().unwrap();
        *count += 1;
        response!("count in myrustlet={}", count);
    });

    // define our second rustlet
    rustlet!("myrustlet2", {
        let mut count = count_clone.lock().unwrap();
        *count += 1;
        response!("count in myrustlet2={}", count);
    });

    // define mappings to our rustlets
    rustlet_mapping!("/myrustlet", "myrustlet");
    rustlet_mapping!("/myrustlet2", "myrustlet2");

    Ok(())
}