Macro librustlet::session

macro_rules! session {
    ($a:expr) => { ... };
    ($a:expr,$b:expr) => { ... };
}

Sets or gets a value in the session. The session is an in-memory key/value store that can be used by rustlets to store data that can be accessed by other rustlets. A session cookie is set called rustletsessionid that lets the rustlet container know which user is which. The session is automatically invalidated after a certain period of time where no calls to session! or session_delete! are made. By default, this amount of time is 30 minutes, but it is configurable in crate::RustletConfig::session_timeout. If only one parameter is specified, the value is retrieved from the session data store, if two parameters are specified, the value is set, see the examples below for more details.

Examples

use nioruntime_util::Error;
use librustlet::*;
use nioruntime_log::*;

debug!();

// implement a simple `Writeable` and `Readable` struct to demonstrate how the session works.
#[derive(Debug)]
struct Example {
   num: u32,
}

impl Example {
    pub fn new(num: u32) -> Self {
        Example { num }
    }
}

impl Writeable for Example {
    fn write<W: Writer>(&self, writer: &mut W) -> Result<(), Error> {
        writer.write_u32(self.num)?;
        Ok(())
    }
}

impl Readable for Example {
    fn read<R: Reader>(reader: &mut R) -> Result<Self, Error> {
        let num = reader.read_u32()?;
        Ok(Example { num })
    }  
}

fn test() -> Result<(), Error> {

    // init the rustlet container, in this case with default values
    rustlet_init!(RustletConfig::default());

    // this rustlet gets the value stored in session variable "abc". If it has not been set
    // (i.e. by the set_session rustlet, 'none' is displayed.
    rustlet!("get_session", {
        let value: Option<Example> = session!("abc"); // the type 'Example' is coerced here
        match value {
            Some(value) => {
                response!("abc={:?}", value); // print out it's value
            }
            None => {
                response!("none"); // print out none
            }
        }
    });
     
    rustlet!("set_session", {
        // get the value of 'abc' in the query string and try to parse as u32
        let val: u32 = query!("abc").parse()?;
        // create an Example with this value and insert it into the session under var 'abc'
        session!("abc", Example::new(val));
    });
     
    // delete the entire session
    rustlet!("delete_session", {
        session_delete!();
    });
     
    // delete only the 'abc' value from the session
    rustlet!("delete_abc", {
        session_delete!("abc");
    });

    rustlet_mapping!("/get_session", "get_session");
    rustlet_mapping!("/set_session", "set_session");
    rustlet_mapping!("/delete_session", "delete_session");
    rustlet_mapping!("/delete_abc", "delete_abc");
     
    Ok(())
}