Macro librustlet::socklet

macro_rules! socklet {
    ($a : expr, $b : expr) => { ... };
}

Specifies a socklet. Socklets are closures that are the equivalent to rustlet’s, except they handle websockets instead of HTTP requests. Since, like rustlets, socklets are closures variables can be moved into them and shared with other socklets and rustlets or any other closure. Socklets are processed in the nioruntime. So, the exectuion is performant. See the other macros for detailed examples on how to use all of the functionality of socklets.

Also see

rustlet_init

handle

socklet_mapping

event

text

binary

ping

pong

Examples

use nioruntime_err::Error;
use librustlet::*;
use nioruntime_log::*;

debug!();

fn test() -> Result<(), Error> {
    rustlet_init!(
        RustletConfig {
            http_config: HttpConfig {
                port: 80,
                ..HttpConfig::default()
            },
            ..RustletConfig::default()
        }
    );

    socklet!("perfsocklet", {
        let handle = handle!()?;
        match event!()? {
            Socklet::Binary => {
                let bin = binary!()?;
                binary!(handle, bin);
            }
            _ => {}
        }
    });

    socklet_mapping!("/perfsocklet", "perfsocklet");

    socklet!("mysocklet", {
        let handle = handle!()?;
        let id = handle.get_connection_id();
        match event!()? {
            Socklet::Open => {
                info!("socklet [cid={}] open!", id);
            }
            Socklet::Close => {
                info!("socklet [cid={}] close!", id);
            }
            Socklet::Text => {
                let text = text!()?;
                info!("got text [cid={}]: {}", id, text);
                text!(handle, "echo [cid={}]: '{}'", id, text,);
            }
            Socklet::Binary => {
                let bin = binary!()?;
                info!("got binary [cid={}]: {:?}", id, bin);
                binary!(handle, [0u8, 1u8, 2u8, 3u8]);
                if bin.len() > 0 && bin[0] == 100 {
                    ping!(handle);
                }
            }
            Socklet::Ping => {
                pong!(handle);
            }
            Socklet::Pong => {}
            }
    });

    socklet_mapping!("/mysocklet", "mysocklet");
    Ok(())
}