Struct librustlet::rustlet_impls::HttpServer

pub struct HttpServer {
    pub config: HttpConfig,
    pub http_context: Option<Arc<RwLock<HttpContext>>>,
    // some fields omitted
}

The main struct representing an HttpServer.

Examples

use nioruntime_http::HttpConfig;
use nioruntime_http::HttpServer;
use nioruntime_err::Error;

fn test() -> Result<(), Error> {
    let config = HttpConfig {
        port: 80,
        ..HttpConfig::default()
    };
    let mut http_server = HttpServer::new(config);
    http_server.start()?;
    Ok(())
}

Fields
config: HttpConfig

The config of this HttpServer.

http_context: Option<Arc<RwLock<HttpContext>>>
Implementations

impl HttpServer

pub fn new(config: HttpConfig) -> HttpServer

Create a new HttpServer instance based on the HttpConfig specified.

pub fn tor_sign(&self, message: &[u8]) -> Result<[u8; 64], Error>

pub fn verify(
&self,
message: &[u8],
pubkey: Option<[u8; 32]>,
signature: [u8; 64]
) -> Result<bool, Error>

pub fn get_tor_pubkey(&self) -> Result<Option<[u8; 32]>, Error>

pub fn add_api_extension(&self, extension: String) -> Result<(), Error>

add an API extension such that any requests that end in this extension will be sent to the the specified HttpConfig::callback. This is used by rsps to map all paths that end in “.rsp” to the rustlet container instead of being processed by the HttpServer.

pub fn add_api_mapping(&self, path: String) -> Result<(), Error>

Add an API mapping such that any requests that have this URI will be sent to the specified HttpConfig::callback instead of being processed by the HttpServer. This is used by rustlet mappings.

pub fn start(&mut self) -> Result<(), Error>

Start the HttpServer. This function will print some startup parameters to the mainlog, which is initially configured to print to both standard output and the mainlog log file location. After the startup parameters are printed, the server has begun and logging only is printed to the mainlog file location.

pub fn stop(&mut self) -> Result<(), Error>

Stop the HttpServer. All resources including threads and file descriptors will be released.

pub fn get_path(config: &HttpConfig, uri: &str) -> Result<String, Error>

Get the path of the file specified based on the the HttpConfig and the uri.

pub fn write_headers(
wh: &WriteHandle,
config: &HttpConfig,
found: bool,
found_404_content: bool,
keep_alive: bool,
additional_headers: Vec<(String, String), Global>,
redirect: Option<String>
) -> Result<(), Error>

Write headers the connection associated with this WriteHandle.

wh - The write handle to use to write.

config - The HttpConfig associated with this connection.

found - Whether or not this file was found.

keep-alive - Whether or not to keep the connection alive after this response is sent.

additional_headers - Additional headers to send with this response.

redirect - Optional redirect for this request.

pub fn build_headers(
config: &HttpConfig,
found: bool,
found_404_content: bool,
keep_alive: bool,
additional_headers: Vec<(String, String), Global>,
redirect: Option<String>,
buf: &mut [u8]
) -> Result<usize, Error>

Build (but do not write) headers based on specified parameters.

config - The HttpConfig associated with this connection.

found - Whether or not this file was found.

keep-alive - Whether or not to keep the connection alive after this response is sent.

additional_headers - Additional headers to send with this response.

redirect - Optional redirect for this request.

buf - The buffer to write the response into. Note: the caller is responsible for

ensuring the capacity of this buffer is sufficient.

pub fn do_house_keeping(
http_context: &Arc<RwLock<HttpContext>>,
http_config: &HttpConfig
) -> Result<bool, Error>