Module librustlet::rustlet_impls

Structs

ConnData
Connection Data used internally It is held in a lock that is used to determine if the thread has panicked.

HttpConfig
The configuration struct for an HttpServer. The Default trait is implemented for this struct so default can be used like below.

HttpServer
The main struct representing an HttpServer.

RustletAsyncContext

RustletConfig
The configuration of the rustlet container.

RustletContainer

RustletRequest

RustletResponse

SessionData

Type Definitions

Rustlet